requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
Pillow==11.0.0
psycopg2-binary==2.9.10
